requirements.txt
kafka-python==2.0.2
Pyspark==3.5.1
pyyaml==6.0
python-dotenv==0.20.0
confluent-kafka